% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.Olink-package.R
\docType{package}
\name{Useful2me-package}
\alias{Useful2me-package}
\title{Olink : A package for Olink proteomics data}
\description{
A collection of R functions to handle Olink proteomics data
}
\author{
Mun-Gwan Hong <\email{mungwan@gmail.com}>
}
\keyword{package}
